fix: four narrow-scope checks in the support code
core.dumps: the set branch returned an unsorted list(o) directly beneath
the comment explaining why unsorted iteration of a hash-backed collection
is non-deterministic. CarrierStore.quarantined is a plain set, and
frozenset is not a set subclass, so the sorted branch written for Topic
never covered it. sorted() for set, list() kept for tuple (derived_from
is a propagation trail whose order is meaningful).

tools/audit_plan.py: the Vietnamese denylist tested membership over whole
[A-Za-z_]\w* identifiers, so snake_case compounds were invisible --
truoc_khi and so_thieu produced no finding while thieu did, even though
snake_case is how a Vietnamese phrase normally reaches an identifier.
Split each identifier on '_' first. The in-code note documented a
different blind spot (words colliding with English), so the claim it made
about coverage was inaccurate; corrected.

topics.distribution: int(statistics.median(...)) truncated a true x.5 and
reported it as an exact integer. The decision gate is 'median > 3 =>
proceed', so a true 3.5 was reported as 3 and would have tripped a STOP
the data never called for. Return the float; both pools measure 4.0, so
no published number moves -- only its printed form (spike updated).

swebench_fetch.fetch: opened the output with 'w' before the download
loop, so a mid-download failure replaced a complete corpus with a
truncated JSONL that still parses; and .get('rows', []) read an error
payload as end-of-split, returning a short count as if it were the whole
dataset. Write to a .part file and os.replace on success; raise when a
payload has no 'rows' key (an EMPTY rows list is still the terminator).
Three new gate-1 tests drive it through a local stub -- no network.

+++ auditgame/topics.py
@@ -26,6 +26,14 @@ def distribution(rows) -> dict:
     hist: dict = {}
     for s in sizes:
         hist[s] = hist.get(s, 0) + 1
+    # The median is returned as a FLOAT, not int()-truncated. The decision gate
+    # in PLAN.md is "median > 3 => proceed, median <= 3 => STOP and tokenize
+    # finer", and statistics.median averages the two middle values on an even n,
+    # so a true 3.5 is a PROCEED that int() would report as "3" -- a STOP the
+    # data never called for. Both pools measure 4.0 today, so nothing already
+    # published moves; the truncation is removed because the gate reads this
+    # number, not because the number is currently wrong. Formatting is the
+    # caller's business.
     return {"n": len(sizes),
-            "median": int(statistics.median(sizes)) if sizes else 0,
+            "median": float(statistics.median(sizes)) if sizes else 0.0,
             "hist": dict(sorted(hist.items()))}
